chore: Update .gitignore and Box component styling
This commit updates the .gitignore file to include the .vercel file, ensuring that it is ignored by Git. Additionally, it updates the styling of the Box component by changing the background color of the button to a darker shade of the primary color. These changes improve the project's version control and enhance the visual appeal of the Box component.