feat: Update loading page component and color scheme
This commit updates the loading page component in the App.jsx file to use a separate LoadingPage component instead of a fallback div. Additionally, it changes the color scheme of the application by updating the black color in the tailwind.config.js file to '#151019'. These changes improve the user experience and enhance the visual appeal of the application.